backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.20
tensorflow==2.19.0
Pillow==11.3.0
numpy==2.0.2
pydantic==2.11.9
pydantic-settings==2.10.1
requests==2.32.3